Moved seriesinfo to seperate file
